refactor: squash tests into one file
